# Conflicts:
#	app/schemas/com.platform.database.AppDatabase/1.json
#	app/src/main/java/com/platform/presenter/MatrimonyMeetFragmentPresenter.java
#	app/src/main/java/com/platform/utility/Util.java
#	app/src/main/java/com/platform/view/fragments/MatrimonyFragment.java
#	app/src/main/java/com/platform/view/fragments/MatrimonyMeetFragment.java
#	app/src/main/res/layout/fragment_matrimony.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/platform/models/SujalamSuphalam/MouDetails.java
@@ -15,7 +15,7 @@ public class MouDetails {
     private List<String> mOUImages = null;
     @SerializedName("date_of_signing")
     @Expose
-    private Integer dateOfSigning;
+    private long dateOfSigning;
     @SerializedName("is_MOU_cancelled")
     @Expose
     private String isMOUCancelled;
@@ -39,11 +39,11 @@ public class MouDetails {
         this.mOUImages = mOUImages;
     }
 
-    public Integer getDateOfSigning() {
+    public long getDateOfSigning() {
         return dateOfSigning;
     }
 
-    public void setDateOfSigning(Integer dateOfSigning) {
+    public void setDateOfSigning(long dateOfSigning) {
         this.dateOfSigning = dateOfSigning;
     }
